Refractor & Code Cleanup, Aggiunto un TODO
